server/requirements.txt
flask
flask-cors
PyJWT[crypto]
cryptography

